pyproject.toml
[tool.poetry]
name = "clients-base"
version = "0.1.0"
description = ""
authors = ["GGlamer <azamat201760@yandex.ru>"]
readme = "README.md"
packages = [{include = "clients_base"}]

[tool.poetry.dependencies]
python = "^3.11"
passlib = "^1.7.4"
fastapi = "^0.89.1"
uvicorn = "^0.20.0"
databases = "^0.7.0"
python-dotenv = "^0.21.0"
asyncpg = "^0.27.0"
psycopg2 = "^2.9.5"
bcrypt = "^4.0.1"
alembic = "^1.9.1"
python-multipart = "^0.0.5"
python-jose = "^3.3.0"
pre-commit = "^2.21.0"
pydantic = {extras = ["email"], version = "^1.10.4"}


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"